* [#1097] feat(LocalvariablesNamesEnhancer): remove generation of parameter names fields because parameter names are available using standard API in Java 8.
