+++ framework/src/play/utils/Java.java
@@ -13,6 +13,7 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
+import java.lang.reflect.Parameter;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -205,7 +206,7 @@ public class Java {
      * Retrieve parameter names of a method
      */
     public static String[] parameterNames(Method method) throws Exception {
-        java.lang.reflect.Parameter[] parameters = method.getParameters();
+        Parameter[] parameters = method.getParameters();
         String[] names = new String[parameters.length];
         for (int i = 0; i < parameters.length; i++) {
             names[i] = parameters[i].getName();
